feat(mocks): implement mock discovery
You can now set the MOCK_SEARCH_EXPRESSION environment varibale to allow
the mockserver to pick up mocks from that glob. The files matched need
to be in Javascript and have a named "mocks" export with an array of
mocks in it. Each mock consists of a unique id and a request function.

BREAKING CHANGE: the environment variable was not necessary before. Set it to
an non-existant path as a default, so that the behaviour does not
change for you

Closes DCOS-21972