feat(tests): add `requestMock` helper
Introduce a  `requestMock` helper to request mocks using nodes `http` module. This helper enables long-running (long-polling) XHR tests as it enables the test to close (abort) connections if the resp. test conditions match or fail - without this changes Jest throws warnings as there would be an async action that is still running, though the test is completed.