feat(tests): add `toRespondWithChunks` matcher
Add a  `toRespondWithChunks` matcher that can be used in conjunction with the `requestMock` helper or native node `http` requests. The matcher compares the received chunks with the provided ones. It wraps the provided values using `Buffer.from` so that there's no need to do so in the tests. The matcher will exit and close the connection as soon as the condition match or fail.

Please note that this is an async matcher thus the tests to be async as well.

This matcher helps with tests that run against potentially long-running connections.